requirements.txt
diff-match-patch==20121119
Django==1.11.10
django-ckeditor==5.4.0
django-import-export==1.0.0
django-js-asset==1.0.0
djangorestframework==3.7.7
et-xmlfile==1.0.1
jdcal==1.3
mod-wsgi==4.5.20
odfpy==1.3.6
openpyxl==2.5.0
Pillow==5.0.0
pytz==2018.3
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0